+++ packages/connect/src/connector/smoldot-light.test.ts
@@ -189,11 +189,11 @@ describe("SmoldotConnect::smoldot-light", () => {
       mockedChain = mockedSmoldotLight.getLatestClient()._getLatestChain()
       expect(mockedChain._addChainOptions.chainSpec).toEqual("fake-ksmcc3-spec")
 
-      await addWellKnownChain(WellKnownChain.rococo_v2)
+      await addWellKnownChain(WellKnownChain.rococo_v2_1)
 
       mockedChain = mockedSmoldotLight.getLatestClient()._getLatestChain()
       expect(mockedChain._addChainOptions.chainSpec).toEqual(
-        "fake-rococo_v2-spec",
+        "fake-rococo_v2_1-spec",
       )
     })